+++ client/flutter/lib/pages/question_index.dart
@@ -1,5 +1,5 @@
 import 'package:WHOFlutter/components/back_arrow.dart';
-import 'package:WHOFlutter/components/question_data.dart';
+import 'package:WHOFlutter/api/question_data.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
 import 'package:url_launcher/url_launcher.dart';
